profiler/sweeper: organize renderer and output plugins
