+++ components/src/dynamo/profiler/sweeper/renderers/direct/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Direct CONFIG_MODIFIERS-backed Sweeper DGD renderer."""
